Extract PageBuilder class from index.php
